finished Test CV project
Hi, my name is Luis Eduardo, i have been working with iOS technologies for more than 5 years. Because the knowledge should never stop, i really love this world of iOS, the posibilities to build something are endless.

i hope that you like my work and it's a pleasure be a candidate for this position. i would like to be part of your team and provide my experience and keep learning.

best regards.